% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{synSample}
\alias{synSample}
\title{Get posteriors}
\usage{
synSample(
  dta,
  method = c("powerp", "mixp"),
  var_grp = "group",
  var_outcome = "Y",
  vague_ab = c(0.5, 0.5),
  grp_hist = NULL,
  grp_cur = NULL,
  weights = NULL,
  ...
)
}
\arguments{
\item{dta}{data frame that contains group identifier and binary outcome}

\item{method}{\code{powerp} for power prior; \code{mixp} for mixture prior}

\item{var_grp}{column name in dta for the group identifier}

\item{var_outcome}{column name in dta for the binary outcome}

\item{vague_ab}{alpha and beta parameter for the vague beta prior. Defaults
are 0.5, 0.5}

\item{grp_hist}{group identifiers for historical studies}

\item{grp_cur}{group identifiers for current study}

\item{weights}{Depending on the method, power parameters or mixture weights.
Length should be the same as grp_hist. All weights are non-negative. For
mixture weights, the sum has to be less than or equal to 1.}

\item{...}{Parameters for STAN sampling}
}
\value{
Posterior samples of theta
}
\description{
Get posteriors
}
\examples{
\dontrun{
dta_1 <- data.frame(group = 1, y = rbinom(100, 1, 0.8))
dta_2 <- data.frame(group = 2, y = rbinom(100, 1, 0.6))
dta_3 <- data.frame(group = 3, y = rbinom(100, 1, 0.7))
dta   <- rbind(dta_1, dta_2, dta_3)

synSample(dta, method = "powerp",
          var_grp = "group", var_outcome = "y",
          vague_ab = c(0.5, 0.5),
          grp_hist = c(1, 2), grp_cur = 3, weights = c(0, 0))

synSample(dta, method = "mixp",
          var_grp = "group", var_outcome = "y",
          vague_ab = c(0.5, 0.5),
          grp_hist = c(1, 2), grp_cur = 3, weights = c(0, 0.9))}


}
